crash fixed

+++ scrabble.cpp
@@ -15,7 +15,7 @@ int hashFun(string &key, int slots = 1000){
     for (int i = 0; i < key.length(); i++)
         sum += pow((int(key[i]) * (i + 1)),2);
     sum = sqrt(sum);
-    return int(slots * (fmod(KA * double(sum), 1)));
+    return int(slots * (fmodl(KA * double(sum), 1)));
 }
 
 void readFile(std::string file,  map<int, int> &hashNumLengths){
